Added fixes. Changed project name to republish
