+++ app/src/androidTest/java/com/purple_paid/purple/ApplicationTest.java
@@ -1,0 +1,15 @@
+package com.purple_paid.purple;
+
+import android.app.Application;
+import android.test.ApplicationTestCase;
+
+/**
+ * <a href="http://d.android.com/tools/testing/testing_android.html">Testing Fundamentals</a>
+ */
+public class ApplicationTest extends ApplicationTestCase<Application>
+{
+    public ApplicationTest()
+    {
+        super(Application.class);
+    }
+}
