--- app/src/androidTest/java/com/twotonestallion/purple_paid/ApplicationTest.java
@@ -1,15 +1,0 @@
-package com.twotonestallion.purple_paid;
-
-import android.app.Application;
-import android.test.ApplicationTestCase;
-
-/**
- * <a href="http://d.android.com/tools/testing/testing_android.html">Testing Fundamentals</a>
- */
-public class ApplicationTest extends ApplicationTestCase<Application>
-{
-    public ApplicationTest()
-    {
-        super(Application.class);
-    }
-}
