+++ app/src/main/AndroidManifest.xml
@@ -1,9 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.twotonestallion.purple_paid"
+    package="com.purple_paid.purple"
     android:versionCode="1"
     android:versionName="1.0">
-    <!--<uses-permission android:name="android.permission.INTERNET" />-->
     <application
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
@@ -11,7 +10,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name="com.twotonestallion.purple_paid.MainActivity"
+            android:name="com.purple_paid.purple.MainActivity"
             android:label="@string/app_name"
             android:theme="@style/AppTheme"
             android:screenOrientation="portrait">
@@ -22,7 +21,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name="com.twotonestallion.purple_paid.Rules"
+            android:name="com.purple_paid.purple.Rules"
             android:label="@string/title_activity_rules" >
         </activity>
     </application>
