--- app/src/main/java/com/twotonestallion/purple_paid/Rank.java
@@ -1,28 +1,0 @@
-package com.twotonestallion.purple_paid;
-
-public enum Rank
-{
-    TWO(2),
-    THREE(3),
-    FOUR(4),
-    FIVE(5),
-    SIX(6),
-    SEVEN(7),
-    EIGHT(8),
-    NINE(9),
-    TEN(10),
-    JACK(11),
-    QUEEN(12),
-    KING(13),
-    ACE(14);
-
-    private int value;
-
-    Rank(int value) {
-        this.value = value;
-    }
-
-    public int getValue() {
-        return value;
-    }
-}
