+++ app/src/main/res/layout/activity_rules.xml
@@ -8,7 +8,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="com.twotonestallion.purple_paid.Rules">
+    tools:context="com.purple_paid.purple.Rules">
 
     <WebView
         android:layout_width="match_parent"
