+++ app/src/main/res/layout/fragment_game.xml
@@ -2,7 +2,7 @@
                 xmlns:tools="http://schemas.android.com/tools"
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
-                tools:context="com.twotonestallion.purple_paid.game">
+                tools:context="com.purple_paid.purple.game">
 
     <LinearLayout
         android:id="@+id/gameLayout"
